refactor(backups): lift BackupSecrets into commons-servers; jobs worker uses it
Move the per-group repo-password Secret store (Kube | in-memory Memory)
from public-server into commons-servers so all three components share one
testable store; public_server::state re-exports it. Add put_password
(server-side apply, create-or-replace) for the rotation loop. The jobs
worker now reads/writes via BackupSecrets instead of a raw kube client —
which also fixes a latent namespace mismatch (the worker read from
CANOPY_NAMESPACE while onboarding created in POD_NAMESPACE; both now use
BackupSecrets' POD_NAMESPACE).

TAM-6877
