+++ crates/commons-servers/src/lib.rs
@@ -13,6 +13,7 @@ use tower_http::{
 use tracing::Span;
 
 pub mod backup_jobs;
+pub mod backup_secrets;
 pub mod device_auth;
 pub mod headers;
 pub mod health;
