+++ crates/public-server/src/state.rs
@@ -1,151 +1,17 @@
-use std::{
-	collections::BTreeMap,
-	sync::{Arc, Mutex},
-};
+#[cfg(feature = "ui")]
+use std::sync::Arc;
 
 use axum::extract::FromRef;
-use commons_errors::{AppError, Result};
+use commons_errors::Result;
 use commons_servers::tailnet_directory::TailnetDirectory;
 use database::Db;
 #[cfg(feature = "ui")]
 use tera::Tera;
 
-/// Env var that forces the in-memory secret store (no cluster needed). Set by the
-/// e2e fixture so the create/reveal paths work against the real binary; tests use
-/// [`BackupSecrets::memory`] directly.
-const MEMORY_ENV: &str = "CANOPY_BACKUP_SECRETS_MEMORY";
-
-type MemoryStore = Arc<Mutex<BTreeMap<String, BTreeMap<String, String>>>>;
-
-/// Reader/writer for the per-group repo-password Secrets. The surface is
-/// deliberately narrow: read one key, create one Secret. Real deployments use
-/// `Kube` (a namespaced [`kube::Client`]); tests and the e2e binary use the
-/// in-memory `Memory` store so the create + read paths are exercised without a
-/// cluster. Shared by the public-server (`/backup-target` reads) and the
-/// private-server (onboarding Secret creation).
-#[derive(Clone)]
-pub enum BackupSecrets {
-	Kube {
-		client: kube::Client,
-		namespace: String,
-	},
-	Memory(MemoryStore),
-}
-
-impl std::fmt::Debug for BackupSecrets {
-	fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-		match self {
-			Self::Kube { namespace, .. } => f
-				.debug_struct("BackupSecrets::Kube")
-				.field("namespace", namespace)
-				.finish_non_exhaustive(),
-			Self::Memory(_) => f.write_str("BackupSecrets::Memory"),
-		}
-	}
-}
-
-impl BackupSecrets {
-	pub fn new(client: kube::Client, namespace: String) -> Self {
-		Self::Kube { client, namespace }
-	}
-
-	/// An in-memory store for tests / the e2e binary.
-	pub fn memory() -> Self {
-		Self::Memory(Arc::new(Mutex::new(BTreeMap::new())))
-	}
-
-	/// Build a secret store from the ambient cluster config (in-cluster the pod's
-	/// service account; locally `~/.kube/config`), reading from the
-	/// `POD_NAMESPACE` namespace (default `canopy`). If `MEMORY_ENV` is set,
-	/// returns the in-memory store instead. Returns `None` (logged) when no
-	/// cluster config is available, so callers degrade to 502 rather than failing
-	/// startup.
-	pub async fn try_default() -> Option<Self> {
-		if std::env::var_os(MEMORY_ENV).is_some() {
-			tracing::warn!("{MEMORY_ENV} set; using in-memory backup Secret store");
-			return Some(Self::memory());
-		}
-		match kube::Client::try_default().await {
-			Ok(client) => Some(Self::Kube {
-				client,
-				namespace: namespace_from_env(),
-			}),
-			Err(err) => {
-				tracing::warn!(error = ?err, "kube client unavailable; backup Secret ops will 502");
-				None
-			}
-		}
-	}
-
-	/// Read one key out of the named Secret. Maps every failure (missing Secret,
-	/// missing key, non-utf8, API error) to [`AppError::Upstream`] so handlers
-	/// return 502 with a generic body.
-	pub async fn read_password(&self, secret_name: &str, key: &str) -> Result<String> {
-		match self {
-			Self::Kube { client, namespace } => {
-				use k8s_openapi::api::core::v1::Secret;
-				use kube::Api;
-
-				let api: Api<Secret> = Api::namespaced(client.clone(), namespace);
-				let secret = api
-					.get(secret_name)
-					.await
-					.map_err(|e| AppError::Upstream(format!("secret get failed: {e}")))?;
-				let data = secret
-					.data
-					.ok_or_else(|| AppError::Upstream("secret has no data".into()))?;
-				let bytes = data
-					.get(key)
-					.ok_or_else(|| AppError::Upstream(format!("secret has no key {key}")))?;
-				String::from_utf8(bytes.0.clone())
-					.map_err(|_| AppError::Upstream("secret value is not valid utf-8".into()))
-			}
-			Self::Memory(store) => store
-				.lock()
-				.unwrap()
-				.get(secret_name)
-				.and_then(|m| m.get(key))
-				.cloned()
-				.ok_or_else(|| AppError::Upstream(format!("secret get failed: {secret_name}"))),
-		}
-	}
-
-	/// Create the named Secret holding `value` under `key`. Maps failures to
-	/// [`AppError::Upstream`]. Used by the private-server onboarding path (Canopy
-	/// owns the repo passphrase Secret for both modes).
-	pub async fn create_password(&self, secret_name: &str, key: &str, value: &str) -> Result<()> {
-		match self {
-			Self::Kube { client, namespace } => {
-				use k8s_openapi::api::core::v1::Secret;
-				use k8s_openapi::apimachinery::pkg::apis::meta::v1::ObjectMeta;
-				use kube::{Api, api::PostParams};
-
-				let api: Api<Secret> = Api::namespaced(client.clone(), namespace);
-				let secret = Secret {
-					metadata: ObjectMeta {
-						name: Some(secret_name.to_string()),
-						..Default::default()
-					},
-					string_data: Some(BTreeMap::from([(key.to_string(), value.to_string())])),
-					..Default::default()
-				};
-				api.create(&PostParams::default(), &secret)
-					.await
-					.map_err(|e| AppError::Upstream(format!("secret create failed: {e}")))?;
-				Ok(())
-			}
-			Self::Memory(store) => {
-				store
-					.lock()
-					.unwrap()
-					.entry(secret_name.to_string())
-					.or_default()
-					.insert(key.to_string(), value.to_string());
-				Ok(())
-			}
-		}
-	}
-}
+/// The per-group repo-password Secret store now lives in `commons-servers`;
+/// re-exported so existing `public_server::state::BackupSecrets` consumers (and
+/// the `AppState.kube` field) keep working.
+pub use commons_servers::backup_secrets::BackupSecrets;
 
 #[derive(Clone, Debug)]
 pub struct AppState {
@@ -245,12 +111,6 @@ impl AppState {
 			kube: None,
 		})
 	}
-}
-
-/// The k8s namespace whose repo-password Secrets `/backup-target` reads, from
-/// `POD_NAMESPACE` (inject via the downward API), defaulting to `canopy`.
-fn namespace_from_env() -> String {
-	std::env::var("POD_NAMESPACE").unwrap_or_else(|_| "canopy".to_string())
 }
 
 impl FromRef<AppState> for crate::ratelimit::RateLimiter {
